+++ src/components/Navbar/Navbar.jsx
@@ -1,31 +1,25 @@
-import React, { useState } from "react";
+import React, { useContext, useState } from "react";
 import { Navbar, Nav, Container, Dropdown } from "react-bootstrap";
-import PropTypes from "prop-types";
 import { FaUserCircle, FaUserAlt, FaPlus } from "react-icons/fa";
-import { IoIosNotifications } from "react-icons/io";
-import { GrMail } from "react-icons/gr";
 import { MdSettings } from "react-icons/md";
 import { FiLogOut } from "react-icons/fi";
 import { useTranslation } from "next-i18next";
+import { auth } from "src/firebase/firebase";
+import userContext from "../../context/context";
 import Searchbox from "./Searchbox";
-import { LoginBtn, JoinBtn, Img, ProfileImg, DropDown } from "./Navbar.styled";
-import LoginModal from "../Modals/LoginModal";
-import SignupModal from "../Modals/SignupModal";
+import { JoinBtn, Img, ProfileImg, DropDown } from "./Navbar.styled";
+import JoinModal from "../Modals/JoinModal";
 
-// eslint-disable-next-line arrow-body-style
-const NavBar = ({ overview, profilePicture, name, uid, handleLogOut }) => {
-  const [showLogin, setShowLogin] = useState(false);
-  const [showSignup, setShowSignup] = useState(false);
+const NavBar = () => {
+  const [showJoin, setShowJoin] = useState(false);
 
-  const openLoginModal = () => {
-    setShowLogin((show) => !show);
-  };
-
-  const openSignupModal = () => {
-    setShowSignup((show) => !show);
+  const openJoinModal = () => {
+    setShowJoin((show) => !show);
   };
 
   const { t } = useTranslation("navbar");
+
+  const user = useContext(userContext);
 
   return (
     <Navbar bg="primary" collapseOnSelect expand="xl" variant="dark">
@@ -44,21 +38,10 @@ const NavBar = ({ overview, profilePicture, name, uid, handleLogOut }) => {
         <Navbar.Collapse id="responsive-navbar-nav">
           <Searchbox />
 
-          {overview ? (
+          {!user ? (
             <Nav className="w-100 d-flex justify-content-end">
-              <LoginBtn onClick={openLoginModal}>{t("navbar.login")}</LoginBtn>
-              <LoginModal
-                showLogin={showLogin}
-                setShowLogin={setShowLogin}
-                setShowSignup={setShowSignup}
-              />
-
-              <JoinBtn onClick={openSignupModal}> {t("navbar.join")} </JoinBtn>
-              <SignupModal
-                showSignup={showSignup}
-                setShowSignup={setShowSignup}
-                setShowLogin={setShowLogin}
-              />
+              <JoinBtn onClick={openJoinModal}>{t("navbar.join")}</JoinBtn>
+              <JoinModal showJoin={showJoin} setShowJoin={setShowJoin} />
             </Nav>
           ) : (
             <Nav className="w-100 d-flex justify-content-end">
@@ -72,20 +55,19 @@ const NavBar = ({ overview, profilePicture, name, uid, handleLogOut }) => {
                 </Dropdown.Toggle>
 
                 <Dropdown.Menu>
-                  <Dropdown.Item href={`./profile/${uid}`}>
-                    {profilePicture ? (
-                      <ProfileImg src={profilePicture} roundedCircle fluid />
+                        {user.photoURL ? (
+                          <ProfileImg src={user.photoURL} roundedCircle fluid />
                     ) : (
                       <FaUserCircle size="28px" />
                     )}
-                    <span className="mx-3 text-muted">{name}</span>
+                        <span className="mx-3 text-muted">{user.displayName}</span>
                   </Dropdown.Item>
 
                   <Dropdown.Item href="/settings">
                     <MdSettings size="20px" />
                     <span className="mx-3 text-muted">{t("navbar.settings")}</span>
                   </Dropdown.Item>
-                  <Dropdown.Item onClick={handleLogOut}>
+                  <Dropdown.Item onClick={() => auth.signOut()}>
                     <FiLogOut size="20px" />
                     <span className="mx-3 text-muted">{t("navbar.logout")}</span>
                   </Dropdown.Item>
@@ -97,20 +79,6 @@ const NavBar = ({ overview, profilePicture, name, uid, handleLogOut }) => {
       </Container>
     </Navbar>
   );
-};
-
-NavBar.propTypes = {
-  overview: PropTypes.bool,
-  profilePicture: PropTypes.string,
-  name: PropTypes.string,
-  uid: PropTypes.string,
-  handleLogOut: PropTypes.func,
-};
-
-NavBar.defaultProps = {
-  overview: false,
-  profilePicture: "",
-  name: "",
 };
 
 export default NavBar;
